move query methods to private in relation module
